# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ARAL System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login for school head or teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Super admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/set-password:
    post:
      operationId: setPassword
      tags: [auth]
      summary: Set initial password for school head first login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPasswordInput"
      responses:
        "200":
          description: Password set successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /grade-levels:
    get:
      operationId: getGradeLevels
      tags: [gradeLevels]
      summary: Get grade levels for this school
      responses:
        "200":
          description: Grade levels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GradeLevel"
    post:
      operationId: createGradeLevel
      tags: [gradeLevels]
      summary: Create a grade level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeLevelInput"
      responses:
        "201":
          description: Grade level created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradeLevel"
  /grade-levels/{gradeLevelId}:
    get:
      operationId: getGradeLevel
      tags: [gradeLevels]
      summary: Get a grade level
      parameters:
        - name: gradeLevelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Grade level
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradeLevel"

  # Teachers
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: ["string", "null"]
      required:
        - error
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
        requiresPasswordChange:
          type: boolean
        profileComplete:
          type: boolean
      required:
        - token
        - user
    AuthUser:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [super_admin, school_head, teacher]
        name:
          type: string
        schoolId:
          type: ["string", "null"]
        schoolName:
          type: ["string", "null"]
        gradeLevelId:
          type: ["string", "null"]
      required:
        - id
        - role
        - name
    LoginInput:
      type: object
      properties:
        schoolId:
          type: string
        credential:
          type: string
        role:
          type: string
          enum: [school_head, teacher]
        teacherId:
          type: ["string", "null"]
      required:
        - schoolId
        - credential
        - role
    AdminLoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    SetPasswordInput:
      type: object
      properties:
        schoolId:
          type: string
        tempCredential:
          type: string
          description: Optional fallback for first-login setup; bearer token is preferred.
        newPassword:
          type: string
      required:
        - schoolId
        - newPassword
...
    GradeLevel:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        name:
          type: string
        teachersCount:
          type: integer
        learnersCount:
          type: integer
        aralLearnersCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - schoolId
        - name
        - teachersCount
        - learnersCount
        - aralLearnersCount
        - createdAt
    GradeLevelInput:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...