# lib/api-zod/src/generated/types/setPasswordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARAL System API
 * OpenAPI spec version: 0.1.0
 */

export interface SetPasswordInput {
  schoolId: string;
  /** Optional fallback for first-login setup; bearer token is preferred. */
  tempCredential?: string;
  newPassword: string;
}